chore: move manager app to auto_validator.core
