--- app/src/manager/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class ManagerConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "manager"
